Changed navbar stuff
